requirements.txt
schedule
python-dotenv
prompt_toolkit